requirements.txt
pdfplumber>=0.11.0
python-dotenv>=1.0.0
langfuse>=4.0.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langgraph>=0.2.0
gradio>=5.0.0
